ai/requirements.txt
kafka-python
river